tests/wpt/tests/screen-wake-lock/wakelock-disabled-by-permissions-policy.https.html

<!DOCTYPE html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/permissions-policy/resources/permissions-policy.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script>
  "use strict";

  const same_origin_src =
    "/permissions-policy/resources/permissions-policy-screen-wakelock.html";
  const cross_origin_src =
    get_host_info().HTTPS_REMOTE_ORIGIN + same_origin_src;

  const header = 'Permissions-Policy header "screen-wake-lock=()"';

  promise_test(t => {
    return promise_rejects_dom(t, "NotAllowedError", navigator.wakeLock.request("screen"));
  }, `${header} disallows the top-level document.`);

  async_test(t => {
    test_feature_availability(
      'navigator.wakeLock.request("screen")',
      t,
      same_origin_src,
      expect_feature_unavailable_default,
      'screen-wake-lock',
    );
  }, `${header} disallows same-origin iframes.`);

  async_test(t => {
    test_feature_availability(
      'navigator.wakeLock.request("screen")',
      t,
      cross_origin_src,
      expect_feature_unavailable_default,
      'screen-wake-lock'
    );
  }, `${header} disallows cross-origin iframes.`);
</script>
</body>
